requirements.txt
Pillow==8.1.2
win10toast==0.9
pypiwin32==223
pywin32==300
setuptools==54.1.1
